requirements.txt
PyGithub==1.55
requests==2.31.0